test coverage

+++ src/main/kotlin/com/kryszak/gwatlin/api/pvp/model/season/SettingsTier.kt
@@ -4,5 +4,5 @@ package com.kryszak.gwatlin.api.pvp.model.season
  * Data model for settings tier object
  */
 data class SettingsTier(
-    val range: List<Int>
+    val range: List<Double>
 )
